fix(#307): use longs instead of ints in Pleroma/Akkoma instance poll configs
Who the hell had the idea to use longs for this >:C

+++ mastodon/src/main/java/org/joinmastodon/android/fragments/ProfileFragment.java
@@ -219,7 +219,7 @@ public class ProfileFragment extends LoaderFragment implements OnBackPressedList
 			if(!isOwnProfile)
 				loadRelationship();
 			else if (isInstanceAkkoma()) {
-				maxFields = getInstance().get().pleroma.metadata.fieldsLimits.maxFields;
+				maxFields = (int) getInstance().get().pleroma.metadata.fieldsLimits.maxFields;
 			}
 		}else{
 			profileAccountID=getArguments().getString("profileAccountID");
